[REEF-245]  Move Network test from REEF.Tests to REEF.Network.Tests project
This PR is add a new test project for Network unit tests and move Network unit test cases to the new project

JIRA: REEF-245. (https://issues.apache.org/jira/browse/REEF-245)

This closes #

Author: Julia Wang  Email: jwang98052@yahoo.com
