+++ lang/cs/Org.Apache.REEF.Tests/Org.Apache.REEF.Tests.csproj
@@ -73,11 +73,6 @@ under the License.
     <Compile Include="Functional\MPI\ScatterReduceTest\ScatterReduceTest.cs" />
     <Compile Include="Functional\MPI\ScatterReduceTest\SlaveTask.cs" />
     <Compile Include="Functional\ReefFunctionalTest.cs" />
-    <Compile Include="Network\BlockingCollectionExtensionTests.cs" />
-    <Compile Include="Network\GroupCommunicationTests.cs" />
-    <Compile Include="Network\GroupCommunicationTreeTopologyTests.cs" />
-    <Compile Include="Network\NameServerTests.cs" />
-    <Compile Include="Network\NetworkServiceTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utility\TestDriverConfigGenerator.cs" />
     <Compile Include="Utility\TestExceptions.cs" />
